# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Spring Playground Backend API
  description: Session-based auth/profile/public HTTP API. This OpenAPI schema is
    the machine-readable contract for frontend integration.
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Local development server
tags:
- name: Auth
  description: Session-based authentication endpoints.
- name: Public
  description: Unauthenticated public endpoints.
- name: Profile
  description: Authenticated profile endpoints.
paths:
  /api/profile/me:
    get:
      tags:
      - Profile
      summary: Get current profile
      description: Returns the profile linked to the current authenticated session.
      operationId: currentProfile
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Current profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Missing or invalid JSESSIONID session cookie.
        "404":
          description: Authenticated user has no linked profile.
      security:
      - sessionCookie: []
    put:
      tags:
      - Profile
      summary: Update current profile
      description: Updates the profile linked to the current authenticated session.
      operationId: updateCurrentProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation failed for the submitted profile payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "200":
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Missing or invalid JSESSIONID session cookie.
        "404":
          description: Authenticated user has no linked profile.
      security:
      - sessionCookie: []
  /api/auth/reset-password:
    post:
      tags:
      - Auth
      summary: Reset password
      description: Resets a password using a valid reset token.
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Reset token is invalid or expired.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Password reset completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResponse"
  /api/auth/register:
    post:
      tags:
      - Auth
      summary: Register
      description: "Creates a new user account and default profile, then sends an\
        \ email verification link when mail delivery succeeds."
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "409":
          description: Username or email is already in use.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Successful registration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
  /api/auth/verify-email:
    post:
      tags:
      - Auth
      summary: Verify email
      description: Confirms a user email using a token that was previously sent by
        email.
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Verification token is invalid or expired.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Email verification completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyEmailResponse"
  /api/auth/resend-verification-email:
    post:
      tags:
      - Auth
      summary: Resend verification email
      description: Accepts an email address and sends a fresh verification link when
        the account exists and is not yet verified. The response is still accepted
        when the account is missing or already verified.
      operationId: resendVerificationEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationEmailRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResendVerificationEmailResponse"
  /api/auth/logout:
    post:
      tags:
      - Auth
      summary: Logout
      description: Invalidates the current server-side session. The endpoint returns
        204 even when no active session exists.
      operationId: logout
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: Logout completed.
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Login
      description: Authenticates by username or email and creates a server-side session
        backed by the JSESSIONID cookie.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Successful login. The response body describes the authenticated
            user and the response sets JSESSIONID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials, rejected login request, or email is not
            yet verified.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/forgot-password:
    post:
      tags:
      - Auth
      summary: Forgot password
      description: Accepts an email address and sends a reset link when the account
        exists. The response is still accepted when the account is missing.
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordResponse"
  /api/public/ping:
    get:
      tags:
      - Public
      summary: Public ping
      description: Returns a simple public status response so callers can verify that
        the backend is reachable.
      operationId: ping
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Public ping response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPingResponse"
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
      - error
    UpdateProfileRequest:
      type: object
      description: Payload used to update the current authenticated user's profile.
      properties:
        displayName:
          type: string
          description: User-facing display name.
          example: Updated Demo
          maxLength: 120
          minLength: 0
        bio:
          type: string
          description: Short profile biography. Defaults to an empty string when omitted.
          example: Updated from frontend
          maxLength: 600
          minLength: 0
      required:
      - displayName
    ProfileResponse:
      type: object
      description: Current authenticated user's profile.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: demo
        email:
          type: string
          example: demo@example.com
        role:
          type: string
          example: USER
        displayName:
          type: string
          example: Demo User
        bio:
          type: string
          example: Session-backed example profile
      required:
      - bio
      - displayName
      - email
      - id
      - role
      - userId
      - username
    ApiErrorResponse:
      type: object
      description: Framework-generated error response body.
      properties:
        timestamp:
          type: string
          example: 2026-03-20T06:22:13.782Z
        status:
          type: integer
          format: int32
          example: 400
        error:
          type: string
          example: Bad Request
        path:
          type: string
          example: /api/profile/me
      required:
      - error
      - path
      - status
      - timestamp
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 100
          minLength: 8
      required:
      - newPassword
      - token
    ResetPasswordResponse:
      type: object
      properties:
        reset:
          type: boolean
      required:
      - reset
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 0
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 100
          minLength: 8
      required:
      - email
      - password
      - username
    RegisterResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
      required:
      - email
      - userId
      - username
    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    VerifyEmailResponse:
      type: object
      properties:
        verified:
          type: boolean
      required:
      - verified
    ResendVerificationEmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 1
      required:
      - email
    ResendVerificationEmailResponse:
      type: object
      properties:
        accepted:
          type: boolean
      required:
      - accepted
    LoginRequest:
      type: object
      description: Credentials used for session-based login.
      properties:
        usernameOrEmail:
          type: string
          description: Username or email address of the user.
          example: demo
          minLength: 1
        password:
          type: string
          description: Plain-text password submitted to the backend for authentication.
          example: demo-password
          minLength: 1
      required:
      - password
      - usernameOrEmail
    LoginResponse:
      type: object
      description: Session login result returned after successful authentication.
      properties:
        authenticated:
          type: boolean
          example: true
        userId:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: demo
        email:
          type: string
          example: demo@example.com
        role:
          type: string
          example: USER
      required:
      - authenticated
      - email
      - role
      - userId
      - username
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
      required:
      - email
    ForgotPasswordResponse:
      type: object
      properties:
        accepted:
          type: boolean
      required:
      - accepted
    PublicPingResponse:
      type: object
      description: Simple public reachability response.
      properties:
        status:
          type: string
          example: ok
      required:
      - status
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: Session cookie returned by POST /api/auth/login and sent on protected
        requests.
      name: JSESSIONID
      in: cookie